+++ tests/ClientTest.php
@@ -11,8 +11,9 @@ use GuzzleHttp\Psr7\Request;
 use GuzzleHttp\Psr7\Response;
 use GuzzleHttp\Psr7\Uri;
 use Psr\Http\Message\ResponseInterface;
+use PHPUnit\Framework\TestCase;
 
-class ClientTest extends \PHPUnit_Framework_TestCase
+class ClientTest extends TestCase
 {
     public function testUsesDefaultHandler()
     {
@@ -200,7 +201,7 @@ class ClientTest extends \PHPUnit_Framework_TestCase
         $handler = HandlerStack::create($mock);
         $client = new Client(['handler' => $handler]);
         $client->get('http://foo.com', ['allow_redirects' => true]);
-        $this->assertInternalType('array',  $mock->getLastOptions()['allow_redirects']);
+        $this->assertInternalType('array', $mock->getLastOptions()['allow_redirects']);
     }
 
     /**
@@ -394,6 +395,18 @@ class ClientTest extends \PHPUnit_Framework_TestCase
         $last = $mock->getLastOptions();
         $this->assertEquals([
             CURLOPT_HTTPAUTH => 2,
+            CURLOPT_USERPWD  => 'a:b'
+        ], $last['curl']);
+    }
+
+    public function testAuthCanBeArrayForNtlmAuth()
+    {
+        $mock = new MockHandler([new Response()]);
+        $client = new Client(['handler' => $mock]);
+        $client->get('http://foo.com', ['auth' => ['a', 'b', 'ntlm']]);
+        $last = $mock->getLastOptions();
+        $this->assertEquals([
+            CURLOPT_HTTPAUTH => 8,
             CURLOPT_USERPWD  => 'a:b'
         ], $last['curl']);
     }
@@ -673,5 +686,13 @@ class ClientTest extends \PHPUnit_Framework_TestCase
             'baz',
             (string) $mockHandler->getLastRequest()->getUri()
         );
+    }
+
+    /**
+     * @expectedException InvalidArgumentException
+     */
+    public function testHandlerIsCallable()
+    {
+        new Client(['handler' => 'not_cllable']);
     }
 }