+++ tests/Handler/MockHandlerTest.php
@@ -6,11 +6,12 @@ use GuzzleHttp\Promise\PromiseInterface;
 use GuzzleHttp\Psr7\Request;
 use GuzzleHttp\Psr7\Response;
 use GuzzleHttp\TransferStats;
+use PHPUnit\Framework\TestCase;
 
 /**
  * @covers \GuzzleHttp\Handler\MockHandler
  */
-class MockHandlerTest extends \PHPUnit_Framework_TestCase
+class MockHandlerTest extends TestCase
 {
     public function testReturnsMockResponse()
     {
@@ -26,6 +27,11 @@ class MockHandlerTest extends \PHPUnit_Framework_TestCase
         $res = new Response();
         $mock = new MockHandler([$res, $res]);
         $this->assertCount(2, $mock);
+    }
+
+    public function testEmptyHandlerIsCountable()
+    {
+        $this->assertCount(0, new MockHandler());
     }
 
     /**
@@ -72,7 +78,7 @@ class MockHandlerTest extends \PHPUnit_Framework_TestCase
         $p->wait();
 
         $this->assertFileExists($filename);
-        $this->assertEquals('TEST CONTENT', file_get_contents($filename));
+        $this->assertStringEqualsFile($filename, 'TEST CONTENT');
 
         unlink($filename);
     }
@@ -88,7 +94,7 @@ class MockHandlerTest extends \PHPUnit_Framework_TestCase
         $p->wait();
 
         $this->assertFileExists($meta['uri']);
-        $this->assertEquals('TEST CONTENT', file_get_contents($meta['uri']));
+        $this->assertStringEqualsFile($meta['uri'], 'TEST CONTENT');
     }
 
     public function testSinkStream()
@@ -101,7 +107,7 @@ class MockHandlerTest extends \PHPUnit_Framework_TestCase
         $p->wait();
 
         $this->assertFileExists($stream->getMetadata('uri'));
-        $this->assertEquals('TEST CONTENT', file_get_contents($stream->getMetadata('uri')));
+        $this->assertStringEqualsFile($stream->getMetadata('uri'), 'TEST CONTENT');
     }
 
     public function testCanEnqueueCallables()
@@ -212,7 +218,7 @@ class MockHandlerTest extends \PHPUnit_Framework_TestCase
         };
         $mock($request, ['on_stats' => $onStats])->wait(false);
         $this->assertSame($e, $stats->getHandlerErrorData());
-        $this->assertSame(null, $stats->getResponse());
+        $this->assertNull($stats->getResponse());
         $this->assertSame($request, $stats->getRequest());
     }
 }