+++ tests/Handler/StreamHandlerTest.php
@@ -7,14 +7,16 @@ use GuzzleHttp\Psr7;
 use GuzzleHttp\Psr7\Request;
 use GuzzleHttp\Psr7\Response;
 use GuzzleHttp\Psr7\FnStream;
+use GuzzleHttp\RequestOptions;
 use GuzzleHttp\Tests\Server;
 use GuzzleHttp\TransferStats;
 use Psr\Http\Message\ResponseInterface;
+use PHPUnit\Framework\TestCase;
 
 /**
  * @covers \GuzzleHttp\Handler\StreamHandler
  */
-class StreamHandlerTest extends \PHPUnit_Framework_TestCase
+class StreamHandlerTest extends TestCase
 {
     private function queueRes()
     {
@@ -48,7 +50,7 @@ class StreamHandlerTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @expectedException \GuzzleHttp\Exception\ConnectException
+     * @expectedException \GuzzleHttp\Exception\RequestException
      */
     public function testAddsErrorToResponse()
     {
@@ -75,7 +77,7 @@ class StreamHandlerTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('8', $response->getHeaderLine('Content-Length'));
         $body = $response->getBody();
         $stream = $body->detach();
-        $this->assertTrue(is_resource($stream));
+        $this->assertInternalType('resource', $stream);
         $this->assertEquals('http', stream_get_meta_data($stream)['wrapper_type']);
         $this->assertEquals('hi there', stream_get_contents($stream));
         fclose($stream);
@@ -269,6 +271,8 @@ class StreamHandlerTest extends \PHPUnit_Framework_TestCase
     public function testAddsProxyByProtocol()
     {
         $url = str_replace('http', 'tcp', Server::$url);
+        // Workaround until #1823 is fixed properly
+        $url = rtrim($url, '/');
         $res = $this->getSendResult(['proxy' => ['http' => $url]]);
         $opts = stream_context_get_options($res->getBody()->detach());
         $this->assertEquals($url, $opts['http']['proxy']);
@@ -320,10 +324,10 @@ class StreamHandlerTest extends \PHPUnit_Framework_TestCase
         $path = $path = \GuzzleHttp\default_ca_bundle();
         $res = $this->getSendResult(['verify' => $path]);
         $opts = stream_context_get_options($res->getBody()->detach());
-        $this->assertEquals(true, $opts['ssl']['verify_peer']);
-        $this->assertEquals(true, $opts['ssl']['verify_peer_name']);
+        $this->assertTrue($opts['ssl']['verify_peer']);
+        $this->assertTrue($opts['ssl']['verify_peer_name']);
         $this->assertEquals($path, $opts['ssl']['cafile']);
-        $this->assertTrue(file_exists($opts['ssl']['cafile']));
+        $this->assertFileExists($opts['ssl']['cafile']);
     }
 
     public function testUsesSystemDefaultBundle()
@@ -653,5 +657,27 @@ class StreamHandlerTest extends \PHPUnit_Framework_TestCase
         $stream = $body->detach();
         $this->assertEquals('hi there... This has a lot of data!', stream_get_contents($stream));
         fclose($stream);
+    }
+
+    public function testHonorsReadTimeout()
+    {
+        Server::flush();
+        $handler = new StreamHandler();
+        $response = $handler(
+            new Request('GET', Server::$url . 'guzzle-server/read-timeout'),
+            [
+                RequestOptions::READ_TIMEOUT => 1,
+                RequestOptions::STREAM => true,
+            ]
+        )->wait();
+        $this->assertEquals(200, $response->getStatusCode());
+        $this->assertEquals('OK', $response->getReasonPhrase());
+        $body = $response->getBody()->detach();
+        $line = fgets($body);
+        $this->assertEquals("sleeping 60 seconds ...\n", $line);
+        $line = fgets($body);
+        $this->assertFalse($line);
+        $this->assertTrue(stream_get_meta_data($body)['timed_out']);
+        $this->assertFalse(feof($body));
     }
 }